refactor: extract generated run schema codecs
